refactor: modularize mock factories
