+++ tests/common/mockFactories/index.js
@@ -1,0 +1,4 @@
+export * from './coreServices.js';
+export * from './loaders.js';
+export * from './entities.js';
+export * from './container.js';
