+++ tests/common/prompting/promptPipelineTestBed.js
@@ -13,7 +13,7 @@ import {
   createMockAIPromptContentProvider,
   createMockPromptBuilder,
   createMockEntity,
-} from '../mockFactories.js';
+} from '../mockFactories';
 import BaseTestBed from '../baseTestBed.js';
 import { describeSuiteWithHooks } from '../describeSuite.js';
