+++ tests/unit/turns/turnManager.advanceTurn.roundStart.test.js
@@ -8,7 +8,7 @@ import {
   SYSTEM_ERROR_OCCURRED_ID,
   TURN_PROCESSING_STARTED,
 } from '../../../src/constants/eventIds.js';
-import { createMockEntity } from '../../common/mockFactories.js';
+import { createMockEntity } from '../../common/mockFactories';
 
 describeTurnManagerSuite(
   'TurnManager: advanceTurn() - Round Start (Queue Empty)',
