+++ tests/unit/turns/turnManager.lifecycle.test.js
@@ -12,7 +12,7 @@ import {
   SYSTEM_ERROR_OCCURRED_ID,
 } from '../../../src/constants/eventIds.js';
 import { beforeEach, expect, jest, test, afterEach } from '@jest/globals';
-import { createMockTurnHandler } from '../../common/mockFactories.js';
+import { createMockTurnHandler } from '../../common/mockFactories';
 import { createAiActor } from '../../common/turns/testActors.js';
 
 // --- Test Setup Helpers ---
